+++ TestTools/StreamChatTestTools/Extensions/Unique/ChatMessage+Unique.swift
@@ -54,6 +54,7 @@ extension ChatMessage {
             poll: nil,
             textUpdatedAt: nil,
             draftReply: nil,
+            reminder: nil,
             sharedLocation: nil
         )
     }