+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## StreamChat
 ### ✅ Added
+- Add new `Filter.isNil` to make it easier to query by nil values [#3623](https://github.com/GetStream/stream-chat-swift/pull/3623)
+- Add Message Reminders [#3623](https://github.com/GetStream/stream-chat-swift/pull/3623)
+   - Add `ChatMessageController.createReminder()`
+   - Add `ChatMessageController.updateReminder()` 
+   - Add `ChatMessageController.deleteReminder()`
+   - Add `MessageReminderListController` and `MessageReminderListQuery`
 - Add `ChatMessageController.partialUpdateMessage()` [#3531](https://github.com/GetStream/stream-chat-swift/pull/3531)
 - Add Location Sharing Support [#3531](https://github.com/GetStream/stream-chat-swift/pull/3531)
   - Add `ChatMessage.sharedLocation`
@@ -19,8 +25,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - `didStartSharingLiveLocation()`
     - `didStopSharingLiveLocation()`
     - `didChangeActiveLiveLocationMessages()`
+
+# [4.79.1](https://github.com/GetStream/stream-chat-swift/releases/tag/4.79.1)
+_June 03, 2025_
+
+## StreamChat
 ### 🐞 Fixed
 - Fix an issue where completion handler was called twice after waiting for token refresh [#3683](https://github.com/GetStream/stream-chat-swift/pull/3683)
+- Fix message not marked as published if it was previously intercepted [#3687](https://github.com/GetStream/stream-chat-swift/pull/3687)
 
 # [4.79.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.79.0)
 _May 28, 2025_